Added API to allow resize of frame after reception
Vector backing receieved frames is set to it's max length(516)
however remote end may send packet with size smaller than 516.
Added API which allows caller to resize the frame to number of
bytes received. This is important for the case of DATA packets
where only receiver callback is aware of bytes received.