Coorects callback invocation on receive timeout
receive timeout gets cancelled if packet have arrived in stipulated
period of time. Cancellation of timer will lead to invocation of
it's callback which in turn calls back to user callback. This is
incorrect since timer hasn't expired on packet reception. This was
leading to multiple invocation of user callback. This is corrected
with this commit